// src/app/api/generated/model/loginUserDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Boards API
 * REST API for Boards, Spring Boot version
 * OpenAPI spec version: 1.0.0
 */

export interface LoginUserDto {
  /** @minLength 1 */
  email: string;
  /**
     * @minLength 4
     * @maxLength 255
     */
  password: string;
}